requirements.txt
fastapi[all]
uvicorn[standard]
SQLAlchemy
asyncpg
pydantic
pydantic-settings
alembic
aiomysql
python-dotenv